Fix boundary condition when looping over columns
When looping over columns in KFast, the algorithm checks that there are at least 32 + 3 (for the border) columns available. However, the correct condition would then be `j<=cols-35', or simpy `j<cols-34`.

To verify this, say we have an image with 35 columns. In the current version, the condition 'j<cols-35' would be '0<0', which is false, so the algorithm would not enter the for loop that uses the "full" version of processCols, although it should, because we have enough columns (columns 0-31 for the vectorized registers, and columns 32, 33, 34 for the border).

Now one could say that this error is harmless, since it will just make `processCols` a tiny bit slower, but that's unfortunately not the case.
Line 159 of process cols has:
  last_cols_mask = (1 << (cols - j - 3)) - 1;
and here, if cols is 35, this will amount to: `(1 << 32) -1`, which will overflow since 1<<32 can't be represented in 32 bits (with gcc the result of this is 1).